+++ Bounding_volumes/include/CGAL/Min_ellipse_2/Optimisation_ellipse_2.h
@@ -24,6 +24,7 @@
 #include <CGAL/Conic_2.h>
 #include <CGAL/Optimisation/assertions.h>
 #include <CGAL/Kernel/global_functions_2.h>
+#include <CGAL/Kernel/Conic_misc.h>
 
 
 namespace CGAL {
