--- Conic_2/package_info/Conic_2/copyright
@@ -1,5 +1,0 @@
-Utrecht University (The Netherlands),
-ETH Zurich (Switzerland),
-INRIA Sophia-Antipolis (France),
-Max-Planck-Institute Saarbruecken (Germany),
-Tel-Aviv University (Israel).
