--- Conic_2/package_info/Conic_2/description.txt
@@ -1,1 +1,0 @@
-This package ....
