--- Conic_2/package_info/Conic_2/license.txt
@@ -1,1 +1,0 @@
-LGPL (v3 or later)
