--- Conic_2/package_info/Conic_2/maintainer
@@ -1,1 +1,0 @@
-Andreas Fabri <Andreas.Fabri@geometryfactory.com>
