+++ Homogeneous_kernel/include/CGAL/Homogeneous/Homogeneous_base.h
@@ -43,7 +43,7 @@
 #include <CGAL/Cartesian/Triangle_2.h>
 #include <CGAL/Homogeneous/VectorH2.h>
 #include <CGAL/Homogeneous/Data_accessorH2.h>
-#include <CGAL/ConicHPA2.h>
+#include <CGAL/Homogeneous/ConicHPA2.h>
 
 #include <CGAL/Homogeneous/Aff_transformationH3.h>
 #include <CGAL/Homogeneous/DirectionH3.h>
