+++ Kernel_23/include/CGAL/Kernel/Conic_misc.h
